requirements.txt
streamlit
keras
tensorflow-cpu
numpy
pillow